:recycle: Refactor destruction views
Split the views for the record manager, reviewer and redirect views into
different files.
