+++ src/archiefvernietigingscomponent/destruction/views/record_manager.py
@@ -1,27 +1,18 @@
 from datetime import date
-from typing import List, Tuple
 
-from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.contrib.messages.views import SuccessMessageMixin
 from django.core.exceptions import PermissionDenied
-from django.db import models, transaction
-from django.db.models import Q
+from django.db import transaction
 from django.http import HttpResponseBadRequest
-from django.shortcuts import get_object_or_404
 from django.urls import reverse, reverse_lazy
 from django.utils import timezone
-from django.utils.timesince import timesince
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import CreateView
-from django.views.generic.base import RedirectView, TemplateView
+from django.views.generic.base import TemplateView
 from django.views.generic.edit import FormView
 
 from django_filters.views import FilterView
-from extra_views import (
-    CreateWithInlinesView,
-    InlineFormSetFactory,
-    UpdateWithInlinesView,
-)
+from extra_views import InlineFormSetFactory, UpdateWithInlinesView
 from furl import furl
 from timeline_logger.models import TimelineLog
 from zds_client.client import ClientError
@@ -32,85 +23,25 @@ from archiefvernietigingscomponent.accounts.mixins import (
 )
 from archiefvernietigingscomponent.notifications.models import Notification
 
-from ..constants import RoleTypeChoices
-from .constants import ListItemStatus, ListStatus, Suggestion
-from .filters import RecordManagerListFilter, ReviewerListFilter
-from .forms import (
+from ..constants import ListItemStatus, ListStatus, Suggestion
+from ..filters import RecordManagerListFilter
+from ..forms import (
     DestructionListForm,
     ListItemForm,
     ReviewCommentForm,
-    ReviewForm,
-    ReviewItemBaseForm,
     ZaakArchiveDetailsForm,
     ZaakUrlForm,
     get_reviewer_choices,
     get_zaaktype_choices,
 )
-from .models import (
+from ..models import (
     ArchiveConfig,
     DestructionList,
     DestructionListItem,
-    DestructionListItemReview,
-    DestructionListReview,
     DestructionListReviewComment,
-    StandardReviewAnswer,
 )
-from .service import fetch_zaak
-from .tasks import process_destruction_list, update_zaak, update_zaken
-
-# Views that route to the appriopriate specialized view
-
-
-class EnterView(LoginRequiredMixin, RedirectView):
-    login_url = reverse_lazy("start-page")
-
-    def get_redirect_url(self, *args, **kwargs):
-        role = self.request.user.role
-        if role and role.can_start_destruction:
-            return reverse("destruction:record-manager-list")
-
-        if role and role.can_review_destruction:
-            return reverse("destruction:reviewer-list")
-
-        if self.request.user.is_superuser:
-            return reverse("audit:audit-trail")
-
-        raise PermissionDenied(self.get_permission_denied_message())
-
-
-class DestructionListRedirectView(LoginRequiredMixin, RedirectView):
-    """
-    Redirect the end user to the appropriate detail view for the DL.
-
-    Authors go to the :class:`DestructionListDetailView`, while reviewers go to the
-    review page. Anyone else gets a 403.
-    """
-
-    def get_redirect_url(self, *args, **kwargs) -> str:
-        qs = DestructionList.objects.select_related("author")
-        destruction_list = get_object_or_404(qs, pk=kwargs["pk"])
-
-        if destruction_list.author == self.request.user:
-            return reverse(
-                "destruction:record-manager-detail", kwargs={"pk": destruction_list.pk}
-            )
-
-        permission_denied = PermissionDenied(self.get_permission_denied_message())
-
-        # not an assignee, you have no business hitting this URL!
-        if not destruction_list.assignees.filter(assignee=self.request.user).exists():
-            raise permission_denied
-
-        # okay, you're an assignee, are you also THE assignee?
-        if destruction_list.assignee == self.request.user:
-            return reverse(
-                "destruction:reviewer-create",
-                kwargs={"destruction_list": destruction_list.id},
-            )
-
-        # all the rest -> denied!
-        raise permission_denied
-
+from ..service import fetch_zaak
+from ..tasks import update_zaak, update_zaken
 
 # Record manager views
 
@@ -400,162 +331,3 @@ class UpdateZaakArchiveDetailsView(SuccessMessageMixin, RoleRequiredMixin, FormV
             return super().form_invalid(form)
 
         return super().form_valid(form)
-
-
-# Reviewer views
-
-
-class ReviewerDestructionListView(RoleRequiredMixin, FilterView):
-    """ data for user who can reviewer destruction lists"""
-
-    role_permission = "can_review_destruction"
-    template_name = "destruction/reviewer_list.html"
-    filterset_class = ReviewerListFilter
-    paginate_by = 20
-
-    def get_queryset(self):
-        user = self.request.user
-        review_status = DestructionListReview.objects.filter(
-            author=user, destruction_list=models.OuterRef("id")
-        ).values("status")
-
-        prefiltered_qs = DestructionList.objects.filter(
-            Q(assignee=user) | Q(reviews__author=user)
-        ).distinct()
-
-        return prefiltered_qs.annotate(
-            review_status=models.Subquery(review_status[:1])
-        ).order_by("-created")
-
-    def get_context_data(self, **kwargs) -> dict:
-        context = super().get_context_data(**kwargs)
-
-        config = ArchiveConfig.get_solo()
-        context.update({"can_download_report": config.destruction_report_downloadable})
-        return context
-
-
-class ReviewItemInline(InlineFormSetFactory):
-    model = DestructionListItemReview
-    form_class = ReviewItemBaseForm
-
-
-class ReviewCreateView(RoleRequiredMixin, UserPassesTestMixin, CreateWithInlinesView):
-    model = DestructionListReview
-    form_class = ReviewForm
-    inlines = [ReviewItemInline]
-    template_name = "destruction/review_create.html"
-    success_url = reverse_lazy("destruction:reviewer-list")
-    role_permission = "can_review_destruction"
-
-    def test_func(self):
-        allowed = super().test_func()
-        if allowed is False:
-            return allowed
-
-        destruction_list = self.get_destruction_list()
-        if not destruction_list.assignees.filter(assignee=self.request.user).exists():
-            return False
-
-        return True
-
-    def get_destruction_list(self):
-        list_id = self.kwargs.get("destruction_list")
-        return get_object_or_404(DestructionList, pk=list_id)
-
-    def get_formatted_standard_review_reasons(self) -> List[Tuple[str, str]]:
-        answers = StandardReviewAnswer.objects.all().order_by("order")
-        return [(answer.reason, answer.reason) for answer in answers]
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        formset = context["inlines"][0]
-
-        destruction_list = self.get_destruction_list()
-        standard_review_choices = self.get_formatted_standard_review_reasons()
-        context.update(
-            {
-                "destruction_list": destruction_list,
-                "destruction_list_json": {
-                    "name": destruction_list.name,
-                    "author": destruction_list.author.get_full_name(),
-                    "created": timesince(destruction_list.created),
-                },
-                "formset_config": {
-                    "prefix": formset.prefix,
-                    **{
-                        field.name: int(field.value())
-                        for field in formset.management_form
-                    },
-                },
-                "show_optional_columns": (
-                    not destruction_list.contains_sensitive_info
-                    or self.request.user.role.type != RoleTypeChoices.archivist
-                ),
-                "standard_review_choices": standard_review_choices,
-            }
-        )
-
-        # If the current reviewer has previously requested changes,
-        # add any comments that the user has made when addressing those changes
-        current_reviewer = self.request.user
-        comment = destruction_list.response_to_reviewer(current_reviewer)
-
-        if comment:
-            comment_author = comment.review.destruction_list.author
-            context["comment_to_review"] = {
-                "author": comment_author.get_full_name(),
-                "text": comment.text,
-            }
-
-        return context
-
-    def form_valid(self, form):
-        form.instance.author = self.request.user
-        form.instance.destruction_list = self.get_destruction_list()
-
-        return super().form_valid(form)
-
-    @transaction.atomic
-    def forms_valid(self, form, inlines):
-        response = super().forms_valid(form, inlines)
-
-        # Get the identificaties of the zaken with suggestions
-        zaken_with_suggestions = []
-        for list_item_form in inlines[0]:
-            suggestion = list_item_form.cleaned_data.get("suggestion")
-            if suggestion:
-                zaken_with_suggestions.append(
-                    list_item_form.cleaned_data["identificatie"]
-                )
-
-        # log review
-        list_review = form.instance
-        TimelineLog.log_from_request(
-            self.request,
-            list_review,
-            template="destruction/logs/review_created.html",
-            n_items=len(zaken_with_suggestions),
-            text=list_review.text,
-            items=zaken_with_suggestions,
-        )
-        # send notification
-        message = _("{author} has reviewed the destruction list.").format(
-            author=list_review.author
-        )
-        Notification.objects.create(
-            destruction_list=list_review.destruction_list,
-            user=list_review.destruction_list.author,
-            message=message,
-        )
-
-        destruction_list = self.get_destruction_list()
-        destruction_list.assign(destruction_list.next_assignee(list_review))
-        destruction_list.save()
-
-        if not destruction_list.assignee:
-            transaction.on_commit(
-                lambda: process_destruction_list.delay(destruction_list.id)
-            )
-
-        return response
